Support @Default for required-key-but-not-required-value behavior
