--- src/main/java/org/skife/config/AutoDefault.java
@@ -1,13 +1,0 @@
-package org.skife.config;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import java.lang.annotation.ElementType;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface AutoDefault
-{
-    String value();
-}
